backend/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
coverage==5.5
cryptography==3.4.7
Deprecated==1.2.12
Django==3.2.5
django-oauth-toolkit==1.5.0
django-treebeard==4.5.1
djangorestframework==3.12.4
idna==3.2
jwcrypto==1.0
oauthlib==3.1.1
Pillow==8.3.1
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
selenium==3.141.0
sentry-sdk==1.3.1
six==1.16.0
sqlparse==0.4.1
stripe==2.60.0
toml==0.10.2
urllib3==1.26.6
wrapt==1.12.1
